requirements.txt
bottle
numpy
scikit-learn
matplotlib
pandas
pyyaml
jsonpickle
plotly
psutil
requests
kaleido